Update header stories
